Quick fix for automatic resizing on touch initiation
On every touch initiation it takes data.scale as 1 due to which the image size unexpectedly changes if we aleady have some css transform values set up. 
The fix would be: For the very first time, transform with data.scale. And in rest of the cases set the scaling relative to previous scale values.